requirement.txt
fastapi
uvicorn
pymupdf
psycopg2-binary
pydantic